pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "roblox-outfit-studio"
version = "1.0.0"
description = "Python rewrite of SlothX Outfit Viewer with full 2D clothing editor and 3D avatar preview"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Outfit Studio Contributors" }
]
keywords = ["roblox", "outfit", "3d", "viewer", "clothing", "editor"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Games/Entertainment",
]
dependencies = [
    "PyQt6>=6.5.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "PyOpenGL>=3.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
dist = [
    "pyinstaller>=6.0.0",
]

[project.scripts]
outfit-studio = "main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/ui/*"]
